Move distribution event resolvers into dedicated modules
