+++ back/boxtribute_server/graph_ql/bindables.py
@@ -21,6 +21,11 @@ from ..core.product_category.queries import query as product_category_query
 from ..core.size_range.fields import size_range
 from ..metrics.fields import metrics
 from ..metrics.queries import query as metrics_query
+from ..mobile_distribution.event.fields import distribution_event
+from ..mobile_distribution.event.mutations import (
+    mutation as distribution_event_mutation,
+)
+from ..mobile_distribution.event.queries import query as distribution_event_query
 from ..mobile_distribution.spot.fields import distribution_spot
 from ..mobile_distribution.spot.mutations import mutation as distribution_spot_mutation
 from ..mobile_distribution.spot.queries import query as distribution_spot_query
@@ -36,6 +41,7 @@ query_types = (
     base_query,
     beneficiary_query,
     distribution_spot_query,
+    distribution_event_query,
     metrics_query,
     transfer_agreement_query,
     organisation_query,
@@ -49,6 +55,7 @@ query_types = (
 mutation_types = (
     beneficiary_mutation,
     distribution_spot_mutation,
+    distribution_event_mutation,
     shipment_mutation,
     tag_mutation,
     transfer_agreement_mutation,
@@ -59,6 +66,7 @@ _object_types = [
     base,
     beneficiary,
     distribution_spot,
+    distribution_event,
     metrics,
     organisation,
     product_category,
@@ -79,7 +87,6 @@ def _register_object_type(name):
 
 # ObjectTypes
 box = _register_object_type("Box")
-distribution_event = _register_object_type("DistributionEvent")
 distribution_events_tracking_group = _register_object_type(
     "DistributionEventsTrackingGroup"
 )
